pyproject.toml
[project]
name = "zoya"
version = "0.2.0"
description = "Zoya - Personal AI Employee (Gold Tier)"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=6.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.2.0",
    "mcp>=1.26.0",
    "requests>=2.31.0",
    "flask>=3.0.0",
    "tweepy>=4.14.0",
    "playwright>=1.58.0",
    "schedule>=1.2.2",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "discord.py>=2.3.0",
    "google-generativeai>=0.8.0",
    "anthropic>=0.21.0",
    "imapclient>=2.8.0",
]

[project.scripts]
zoya-watcher = "src.watcher:main"
zoya-orchestrator = "src.orchestrator:main"
zoya-gmail = "src.watchers.gmail_watcher:main"
zoya-whatsapp = "src.watchers.whatsapp_watcher:main"
zoya-linkedin = "src.linkedin_poster:main"
zoya-email-mcp = "src.mcp.email_server:main"
zoya-briefing = "src.briefing_generator:main"
zoya-twitter = "src.twitter_poster:main"
zoya-twitter-mcp = "src.mcp_servers.twitter_mcp:main"
zoya-odoo-mcp = "src.mcp_servers.odoo_mcp:main"
zoya-cross-domain = "src.cross_domain_orchestrator:main"
zoya-audit = "src.audit_generator:main"
zoya-ralph = "src.ralph_wiggum:main"
zoya-bank = "src.watchers.bank_watcher:main"
zoya-smart-reply = "src.automations.smart_reply:main"
zoya-log-janitor = "src.log_janitor:main"
zoya-social-daemon = "src.social_media_daemon:main"
zoya-todo = "src.todo_manager:main"
zoya-email-draft = "src.automations.email_drafter:main"
zoya-research = "src.research_agent:main"
zoya-discord = "src.watchers.discord_watcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
